requirements.txt
gurux-dlms
beautifulsoup4
lxml
pyserial
cryptography
pycryptodomex